+++ src/main/java/com/app/businessBridge/domain/grade/controller/ApiV1GradeController.java
@@ -4,9 +4,11 @@ import com.app.businessBridge.domain.grade.entity.Grade;
 import com.app.businessBridge.domain.grade.request.GradeRequest;
 import com.app.businessBridge.domain.grade.response.GradeResponse;
 import com.app.businessBridge.domain.grade.service.GradeService;
+import com.app.businessBridge.global.RsData.RsCode;
 import com.app.businessBridge.global.RsData.RsData;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,7 +21,11 @@ public class ApiV1GradeController {
 
     // 직급 등록
     @PostMapping("")
-    public RsData post(@Valid @RequestBody GradeRequest.CreateRequest createRequest) {
+    public RsData post(@Valid @RequestBody GradeRequest.CreateRequest createRequest, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return RsData.of(RsCode.F_10, "알 수 없는 오류로 실패했습니다.");
+        }
+
         RsData rsData = this.gradeService.create(createRequest.getGradeCode(),
                 createRequest.getGradeName());
 
@@ -37,7 +43,12 @@ public class ApiV1GradeController {
 
     // 직급 수정
     @PatchMapping("/{id}")
-    public RsData<GradeResponse.PatchedGrade> patch(@Valid @RequestBody GradeRequest.UpdateRequest updateRequest) {
+    public RsData<GradeResponse.PatchedGrade> patch(@Valid @RequestBody GradeRequest.UpdateRequest updateRequest,
+                                                    BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return RsData.of(RsCode.F_10, "알 수 없는 오류로 실패했습니다.");
+        }
+
         RsData<Grade> rsData = this.gradeService.update(updateRequest.getId(), updateRequest.getGradeCode(),
                 updateRequest.getGradeName());
 
@@ -46,7 +57,12 @@ public class ApiV1GradeController {
 
     // 직급 삭제
     @DeleteMapping("/{id}")
-    public RsData delete(@Valid @RequestBody GradeRequest.DeleteRequest deleteRequest) {
+    public RsData delete(@Valid @RequestBody GradeRequest.DeleteRequest deleteRequest,
+                         BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return RsData.of(RsCode.F_10, "알 수 없는 오류로 실패했습니다.");
+        }
+
         RsData rsData = this.gradeService.delete(deleteRequest.getId());
 
         return RsData.of(rsData.getRsCode(), rsData.getMsg());
